Generate strongly-typed instances, when necessary
The commit modifies imxrtral.py to conditionally emit a numbered
peripheral instance. If a peripheral has multiple instances, like
LPUART, GPIO, ENET (1060), we generate an Instance struct that
includes a typenum 'constant.' The constant matches the instance
number; LPUART5's RAL instance is now Instance<typenum::U5>. This is
a breaking change, since before this commit, all peripheral instances
shared the same Instance type, even though they represented a
unique peripheral instance.

This is the effectively imxrt-async-hal's [instance module][1]
represented in the RAL. Additionally, the approach more closely
mimics the svd2rust auto-generated code, which distinguishes
peripheral instances in the type system. I'd like to take advantage
of this new API in both the normal HAL, and the async HAL. It should
pair very well with the imxrt-iomuxc APIs, which also use the typenum
instance numbers.

Let me know if there might be better ways to express this goal in the
Python script. This is my first deep study of the script, so I might
have missed some better implementations.

The next commit will bulk-add all of the auto-generated Rust code.
Study that commit to understand the new code this script generates.

[1]: https://github.com/imxrt-rs/imxrt-async-hal/blob/be400de3159e627061c7a131439e7e769201175c/src/instance.rs#L1